- Moved CollectionView UICollection layout creation to the standalone fabric
