+++ FakeNFT/Scenes/Collection/CollectionDetails/CollectionDetailsView.swift
@@ -11,7 +11,9 @@ final class CollectionDetailsView: UIViewController {
     
     private let collectionViewModel: CollectionDetailsViewModel
     private lazy var collectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: createLayout())
+        let collectionView = UICollectionView(
+            frame: .zero, collectionViewLayout: CollectionViewLayoutFactory.createLayout()
+        )
         return collectionView
     }()
     
@@ -128,7 +130,7 @@ final class CollectionDetailsView: UIViewController {
             stack.translatesAutoresizingMaskIntoConstraints = false
             return stack
         }()
-    
+        
         // MARK: - Collection View Setup
         
         collectionView.backgroundColor = .clear
@@ -181,94 +183,8 @@ final class CollectionDetailsView: UIViewController {
             ]
         )
     }
-    
-    // MARK: -  UICollectionView layout setup
-    
-    private func createLayout() -> UICollectionViewLayout {
-        
-        let layout = UICollectionViewCompositionalLayout { (sectionIndex, layoutEnvironment) -> NSCollectionLayoutSection? in
-            
-            guard let sectionType = Section(rawValue: sectionIndex) else {
-                fatalError("Invalid section")
-            }
-            
-            let inset: CGFloat = 16
-            
-            switch sectionType {
-        
-            // Collection cover image
-            case .cover:
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1.0),
-                    heightDimension: .fractionalWidth(0.83)
-                )
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                let group = NSCollectionLayoutGroup.horizontal(
-                    layoutSize: itemSize,
-                    subitem: item,
-                    count: 1
-                )
-                let section = NSCollectionLayoutSection(group: group)
-                return section
-            
-            // Collection text description
-            case .description:
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1.0),
-                    heightDimension: .estimated(100)
-                )
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                let group = NSCollectionLayoutGroup.horizontal(
-                    layoutSize: itemSize,
-                    subitem: item,
-                    count: 1
-                )
-                
-                let section = NSCollectionLayoutSection(group: group)
-                section.contentInsets = NSDirectionalEdgeInsets(
-                    top: inset,
-                    leading: inset,
-                    bottom: 21,
-                    trailing: inset
-                )
-                
-                return section
-                
-            // Collection NFT list
-            case .collection:
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1/3),
-                    heightDimension: .estimated(192)
-                )
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                let groupSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1.0),
-                    heightDimension: .estimated(192)
-                )
-                let group = NSCollectionLayoutGroup.horizontal(
-                    layoutSize: groupSize,
-                    subitem: item,
-                    count: 3
-                )
-                group.interItemSpacing = .fixed(10)
-                
-                let section = NSCollectionLayoutSection(group: group)
-                section.contentInsets = NSDirectionalEdgeInsets(
-                    top: 0,
-                    leading: inset,
-                    bottom: 0,
-                    trailing: inset
-                )
-                // spacer size + interGroupSpacing
-                section.interGroupSpacing = 20 + 8
-                
-                return section
-            }
-        }
-        
-        return layout
-    }
 }
+
 
 // MARK: - Delegate
